moved ScoreWidget to separate file, added support for different score types
