+++ lib/widgets/module/module_widget.dart
@@ -9,6 +9,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:hearbat/utils/google_tts_util.dart';
 import 'word_missed_button_widget.dart';
 import 'package:confetti/confetti.dart';
+import 'score_widget.dart';
 
 class ModuleWidget extends StatefulWidget {
   final String title;
@@ -293,7 +294,8 @@ class _ModulePageState extends State<ModuleWidget> {
             // Today's score
             ScoreWidget(
               context: context,
-              correctAnswersCount: correctAnswersCount,
+              type: ScoreType.score,
+              correctAnswersCount: correctAnswersCount.toString(),
               subtitleText: "Score",
               icon: Icon(
                 Icons.star,
@@ -305,7 +307,8 @@ class _ModulePageState extends State<ModuleWidget> {
             ),
             ScoreWidget(
               context: context,
-              correctAnswersCount: correctAnswersCount,
+              type: ScoreType.score,
+              correctAnswersCount: correctAnswersCount.toString(),
               subtitleText: "Highest Score",
               icon: Icon(
                 Icons.emoji_events,
@@ -348,113 +351,6 @@ class _ModulePageState extends State<ModuleWidget> {
           ],
         ),
       ],
-    );
-  }
-}
-
-var gradientBoxDecoration = BoxDecoration(
-  gradient: LinearGradient(
-    begin: Alignment.topCenter,
-    end: Alignment.bottomCenter,
-    colors: [
-      Color.fromARGB(255, 248, 213, 245),
-      Color.fromARGB(255, 255, 192, 199),
-      Color.fromARGB(255, 213, 177, 239),
-    ],
-  ),
-  borderRadius: BorderRadius.circular(8.0),
-  border: Border.all(
-    color: Color.fromARGB(255, 7, 45, 78),
-    width: 3.0,
-  ),
-  boxShadow: [
-    BoxShadow(
-      color: Colors.grey.withAlpha((0.5 * 255).toInt()),
-      spreadRadius: 5,
-      blurRadius: 7,
-      offset: Offset(0, 3),
-    ),
-  ],
-);
-
-var blueBoxDecoration = BoxDecoration(
-  color: Color.fromARGB(255, 7, 45, 78),
-  borderRadius: BorderRadius.circular(8.0),
-  border: Border.all(
-    color: Color.fromARGB(255, 7, 45, 78),
-    width: 3.0,
-  ),
-  boxShadow: [
-    BoxShadow(
-      color: Colors.grey.withAlpha((0.5 * 255).toInt()),
-      spreadRadius: 5,
-      blurRadius: 7,
-      offset: Offset(0, 3),
-    ),
-  ],
-);
-
-
-class ScoreWidget extends StatelessWidget {
-  const ScoreWidget(
-      {super.key,
-      required this.context,
-      required this.correctAnswersCount,
-      required this.subtitleText,
-      required this.icon,
-      required this.boxDecoration,
-      required this.total});
-
-  final BuildContext context;
-  final int correctAnswersCount;
-  final String subtitleText;
-  final Icon icon;
-  final BoxDecoration boxDecoration;
-  final int total;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: MediaQuery.of(context).size.width * .90,
-      height: MediaQuery.of(context).size.height * .1,
-      margin: EdgeInsets.fromLTRB(20.0, 20.0, 20.0, 10.0),
-      padding: EdgeInsets.all(8.0),
-      decoration: boxDecoration,
-      child: Stack(
-        children: [
-          Positioned(
-            left: 10,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                AutoSizeText(
-                  '$correctAnswersCount / $total',
-                  style: TextStyle(
-                      fontSize: 25,
-                      fontWeight: FontWeight.w900,
-                      color: subtitleText == "Highest Score"
-                          ? Colors.white
-                          : Color.fromARGB(255, 7, 45, 78)),
-                ),
-                AutoSizeText(
-                  subtitleText,
-                  style: TextStyle(
-                      fontSize: 16,
-                      fontWeight: FontWeight.w900,
-                      color: subtitleText == "Highest Score"
-                          ? Colors.white
-                          : Color.fromARGB(255, 7, 45, 78)),
-                ),
-              ],
-            ),
-          ),
-          Positioned(
-            top: 10,
-            right: 10,
-            child: icon,
-          ),
-        ],
-      ),
     );
   }
 }
